homework3 fixed
