+++ src/main/java/elements/LeftMenu.java
@@ -1,4 +1,4 @@
-package hw3;
+package elements;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -7,25 +7,26 @@ import org.openqa.selenium.support.PageFactory;
 
 import java.util.List;
 
-public class HeaderMenu {
+public class LeftMenu {
 
-    @FindBy(css = "nav[role = 'navigation']>ul>li")
-    private List<WebElement> navigationBarItems;
+    @FindBy(css = "ul[class = 'sidebar-menu left']>li>a>span")
+    private List<WebElement> leftMenuItems;
 
-    public HeaderMenu(WebDriver driver) {
+    public LeftMenu(WebDriver driver) {
         PageFactory.initElements(driver, this);
     }
 
-    public List<WebElement> getNavigationBarItems() {
-        return this.navigationBarItems;
+    public List<WebElement> getLeftMenuItems() {
+        return this.leftMenuItems;
     }
 
-    public void clickNavigatiounBarItem(final String item) {
-        for (WebElement element : navigationBarItems) {
+    public void clickLeftMenuItems(final String item) {
+        for (WebElement element : leftMenuItems) {
             if (item.equals(element.getText())) {
                 element.click();
                 break;
             }
         }
+
     }
 }
